feat(forgetpassowrd | navigation): adds forget password screen - UI + functionalitites
1. User enters the email that is registred with kenaf. If not registred, the user will be navigated
to Register screen to Register first through an alert 2. 5 input boxes will be dispayed depnding on
if the user is registed or not. 3. User will be navigated to a UpdatePassword scrren if the entered
OTP matches the one send on his email and an alert will be shown if the OTP fails