feat(updatepassword): adds input - Password & confirm passoword for sending the password to backend
User is prompted to enter passoword and confirm it by re-entering it again int he conform password
text input. Both the text inputs have validations for empty values , character length and matching.
After the user clicks on the confirm button, the email and password will be send to backend and user
will be navigated to Login screen