fix(verifyotp): stores all the OTP in single state so that the latest one will only be validated
Bug - Earlier there were two states - one for storing the OTP for the first time second was one for
the OTP that is being received after user clicks on the "Didn't get OTP. Click here to send again".
But the condition should be that only the latest OTP should be validated. Solution : Setting the
regeneratedOtp state's initial value to route.params.otp. And when the use cliks on the  Didn't get
OTP. Click here