feat(camerascreen | seperateinvoice - invoice image): maintains width, height while displaying image
This commit maintains the width and height of the image that is being selected from the gallery.
Earlier the image was resized to a fixed width and height, stretching the image to fit the width and
height. Now the selected image will be displayed according to its orignal height and width. The only
issue is the quality of the image is not degraded even when the compression is set to 1 for both the
resizing methods